Convert api context to a class
This allows each library to have its own React context
