--- src/api-ctx/context.ts
@@ -1,5 +1,0 @@
-import {createContext} from "react";
-import ApiContextType from "./ApiContextType";
-
-const ApiContext = createContext<ApiContextType | null>(null);
-export default ApiContext;
