+++ src/api-ctx/index.ts
@@ -1,4 +1,4 @@
-export * from "./ApiProvider";
 export type {default as ApiContextType} from "./ApiContextType";
-export {default as ApiContext} from "./context";
-export {default as useApiContext} from "./hook";
+export {default as createContext} from "./create";
+export * from "./ApiContext";
+export type {default} from "./ApiContext";
