+++ src/index.ts
@@ -1,6 +1,6 @@
 export {default as key} from "./key";
 export type {Param, ParamRecord} from "./key";
-export {ApiProvider} from "./api-ctx";
+export {createContext} from "./api-ctx";
 export type {ApiProviderProps} from "./api-ctx";
 export * from "./swr";
 export type {SWRConfiguration, SWRResponse} from "swr";
